changed async_generator() function typing
changed Generator primary value from int to float

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -6,11 +6,11 @@ from random import uniform
 from typing import Generator
 
 
-async def async_generator() -> Generator[int, None, None]:
+async def async_generator() -> Generator[float, None, None]:
     """[summary]
 
     Yields:
-        Generator[int, None, None]: [description]
+        Generator[float, None, None]: [description]
     """
     for i in range(10):
         yield uniform(0, 10)
